Make rand sample test both wrapper and asktell
