+++ libensemble/gen_classes/gpCAM.py
@@ -34,7 +34,8 @@ class GP_CAM(Generator):
     It is a batched method that produces a first batch uniformly random from
     (lb, ub). On subequent iterations, it calls an optimization method to
     produce the next batch of points. This optimization might be too slow
-    (relative to the simulation evaluation time) for some use cases."""
+    (relative to the simulation evaluation time) for some use cases.
+    """
 
     def _initialize_gpcAM(self, user_specs):
         """Extract user params"""
