+++ libensemble/gen_funcs/persistent_gen_wrapper.py
@@ -9,7 +9,6 @@ def persistent_gen_f(H, persis_info, gen_specs, libE_info):
     ps = PersistentSupport(libE_info, EVAL_GEN_TAG)
     U = gen_specs["user"]
     b = U.get("initial_batch_size") or U.get("batch_size")
-    calc_in = None
 
     generator = U["generator"]
     if inspect.isclass(generator):
@@ -18,6 +17,7 @@ def persistent_gen_f(H, persis_info, gen_specs, libE_info):
         gen = generator
 
     tag = None
+    calc_in = None
     while tag not in [STOP_TAG, PERSIS_STOP]:
         H_o = gen.ask(b)
         tag, Work, calc_in = ps.send_recv(H_o)
